Made new lists for files
